backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
sqlalchemy==1.4.48
aiosqlite==0.18.0
pydantic==1.10.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
googlemaps==4.10.0
polyline==2.0.2
aiohttp==3.8.5